backend/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
firebase-admin==6.6.0
python-dotenv==1.0.1
httpx==0.28.1
